Move test to test_assumptions_2.py
